[fips scan] update log, error message

+++ doozer/doozerlib/cli/scan_fips.py
@@ -24,7 +24,7 @@ class ScanFipsCli:
     @limit_concurrency(os.cpu_count())
     async def run_get_problem_nvrs(self, build: tuple):
         # registry-proxy.engineering.redhat.com/rh-osbs/openshift-ose-sriov-network-operator@sha256:da95750d31cb1b9539f664d2d6255727fa8d648e93150ae92ed84a9e993753be
-        pull_spec = build[1]
+        nvr, pull_spec = build
 
         rc_scan, out_scan, _ = await cmd_gather_async(f"sudo check-payload scan image --spec {pull_spec}")
 
@@ -35,17 +35,25 @@ class ScanFipsCli:
         # c706f2c4 registry-proxy.engineering.redhat.com/rh-osbs/openshift-ose-aws-pod-identity-webhook
         rc, out, err = cmd_gather("sudo podman images --format '{{.ID}} {{.Repository}}'")
         if rc != 0:
-            raise Exception(err)
+            # This does not have a FIPS vulnerability, but we need to figure out why this failing to clean
+            # since memory is limited in buildvm
+            return build
 
         # `out` has multiple lines in {{.ID}} {{.Repository}} format
         for image in out.strip().split("\n"):
+            if not image:
+                # Skip null values
+                continue
+
             image_id, image_name = image.split(" ")  # c706f2c4, registry-proxy.engineering.redhat.com/rh-osbs/openshift-ose-aws-pod-identity-webhook
             if name_without_sha == image_name:
                 self.runtime.logger.info(f"Trying to clean image {image}")
                 rmi_cmd_rc, _, rmi_cmd_err = cmd_gather(f"sudo podman rmi {image_id}")
 
                 if rmi_cmd_rc != 0:
-                    raise Exception(rmi_cmd_err)
+                    # This does not have a FIPS vulnerability, but we need to figure out why this failing to clean
+                    # since memory is limited in buildvm
+                    return build
 
         # The command will fail if it's not run on root, so need to make sure of that first during debugging
         # If it says successful run, it means that the command ran correctly
